+++ src/content/components/ActiveRSMessages/TargetingParserTypes.ts
@@ -1,0 +1,9 @@
+export type Token = {
+  value: string;
+  type: string;
+};
+
+export type ParsedTargetingExpression = {
+  version: any;
+  locale: any;
+};
