+++ src/content/components/IterationPicker/IterationPickerTypes.tsx
@@ -1,0 +1,13 @@
+export interface IterationPickerProps {
+  iterations: string[];
+  currentIteration: string;
+  match: {
+    params: { [key: string]: string };
+    isExact: boolean;
+    path: string;
+    url: string;
+  };
+  history: {
+    push(location: string, state?: unknown): void;
+  };
+}
