+++ src/server/queryUtils.ts
@@ -1,5 +1,13 @@
 import * as request from "request";
-import { IterationLookup, lookupIterations } from "../common/IterationLookup";
+import { lookupIterations } from "../common/IterationLookup";
+import { IterationLookup } from "../common/IterationLookupTypes";
+import {
+  FieldsResponse,
+  QueryConfig,
+  QueryProperties,
+  QueryRuleSet,
+  RSMessage,
+} from "./queryUtilsTypes";
 
 const BZ_BASE_URI = "https://bugzilla.mozilla.org/rest";
 const BZ_BUG_URI = `${BZ_BASE_URI}/bug`;
@@ -8,85 +16,6 @@ const BZ_FIELDS_URI = `${BZ_BASE_URI}/field/bug`;
 // Details about Iterations field
 export const ITERATION_FIELD_NAME = "cf_fx_iteration";
 const BZ_ITERATIONS_URI = `${BZ_FIELDS_URI}/${ITERATION_FIELD_NAME}`;
-
-interface QueryProperties {
-  custom?: Object;
-  operator?: string;
-  key?: string;
-  value?: string;
-  iteration?: string;
-  rules?: QueryProperties | Array<QueryProperties>;
-  include_fields?: Array<string>;
-}
-
-interface QueryConfig extends QueryProperties {
-  rules: QueryProperties | Array<QueryProperties>;
-  include_fields: Array<string>;
-}
-
-type QueryRuleSet =
-  | QueryConfig
-  | Array<QueryConfig>
-  | QueryProperties
-  | Array<QueryProperties>;
-
-interface Message {
-  bugzillaId: string;
-  status: string | React.ReactNode;
-}
-
-export interface RSMessage extends Message {
-  id: string;
-  template: string;
-  targeting: string;
-  parsedTargetingExpression: any;
-  frequency: { lifetime: number };
-  content: any;
-}
-
-interface Field {
-  id: number;
-  type:
-    | 0 // Field type unknown
-    | 1 // Single-line string field
-    | 2 // Single value field
-    | 3 // Multiple value field
-    | 4 // Multi-line text value
-    | 5 // Date field with time
-    | 6 // Bug ID field
-    | 7 // See Also field
-    | 8 // Keywords field
-    | 9 // Date field
-    | 10; // Integer field
-  is_custom: boolean;
-  name: string;
-  display_name: string;
-  is_mandatory: boolean;
-  is_on_bug_entry: boolean;
-  visibility_field: string | null;
-  visibility_values: string[];
-  value_field: string | null;
-  values: Array<FieldValue | KeywordValue | BugStatusValue>;
-}
-
-interface FieldValue {
-  name: string;
-  sort_key: number | void;
-  sortkey?: number | void;
-  visibility_values?: string[];
-  is_active?: boolean;
-}
-
-interface KeywordValue extends FieldValue {
-  description: string;
-}
-
-interface BugStatusValue extends FieldValue {
-  is_open: boolean;
-  can_change_to: { name: string; comment_required: boolean }[];
-}
-
-type FieldsResponse = { fields: Field[] };
 
 // IN PROGRESS
 function _checkGroupOperator(o) {
